@@ -6,7 +6,7 @@
 import 'package:flutter/material.dart';
 
 class RouteUnknown extends StatelessWidget {
-  const RouteUnknown({Key key}) : super(key: key);
+  const RouteUnknown({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
