@@ -3,7 +3,7 @@ import 'package:graph_translator/widgets/widget_graph.dart';
 
 class WidgetZoom extends StatelessWidget {
   final GraphController controller;
-  const WidgetZoom({@required this.controller, Key key}) : super(key: key);
+  const WidgetZoom({required this.controller, Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
@@ -12,7 +12,7 @@ class WidgetZoom extends StatelessWidget {
         IconButton(
           onPressed: () {
             controller.updateViewState((v) {
-              v.applyZoom(1.1);
+              return v..applyZoom(1.1);
             });
           },
           icon: Icon(Icons.zoom_in),
@@ -20,7 +20,7 @@ class WidgetZoom extends StatelessWidget {
         IconButton(
           onPressed: () {
             controller.updateViewState((v) {
-              v.applyZoom(1.0 / 1.1);
+              return v..applyZoom(1.0 / 1.1);
             });
           },
           icon: Icon(Icons.zoom_out),
